backend/pyproject.toml
[project]
name = "wx-massage"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "aiosqlite",
    "pycryptodome",
    "httpx",
]
